fix(pg): widen rate_limits.window_start_ms to BIGINT and add resilience
P0: Rate limits integer overflow
- window_start_ms stored Date.now() (~1.7T) in INTEGER (max 2.1B)
- All rate-limited endpoints crashed with "out of range" error
- Migration 035 changes column to BIGINT
- consumeRateLimit now catches DB errors and returns 503

P1: Missing await in story.js
- writer.getStoryContext() at line 691 was not awaited
- 4 consumeRateLimit calls also missing await

Co-authored by Ambrose Obimma

+++ migrations/pg/001_init.sql
@@ -160,7 +160,7 @@ CREATE TABLE IF NOT EXISTS billing_holds (
 CREATE TABLE IF NOT EXISTS rate_limits (
   user_id TEXT NOT NULL,
   action_type TEXT NOT NULL,
-  window_start_ms INTEGER NOT NULL,
+  window_start_ms BIGINT NOT NULL,
   window_seconds INTEGER NOT NULL,
   count INTEGER NOT NULL,
   limit_count INTEGER NOT NULL,
